refactor(scripts): share the GIR scalar reader
`propsBodies`, `scalarProps` and `tagGTypes` lived inside
`check-adwaita-element-properties.mjs`, and a second ratchet for the
NativeScript surface needs exactly the same question asked of the same
file. Two definitions of "a scalar property" is two backlogs that can
disagree about what they are counting while both stay green — the
SECOND copy is where you lift.

`scripts/gir-scalar-properties.mjs` owns the readers, the exclusions
with their incidents, the fixture and `girReaderSelfTest()`, which both
callers now run: a shared reader only one of its two callers proves is
one half the repository trusts on somebody else's word.

`scalarPropertyNames` is new beside `scalarProps`, returning
`kebab -> camel`. Both spellings are READ from the generated source
rather than mapped back from one another: an HTML attribute is
`can-shrink` and a NativeScript setter is `canShrink`, the generator
emits the pair, and a round-trip that holds for `imModule` today
guarantees nothing about the next generated name. Its own vector is in
the shared self-test.

A/B, each mutation separately with real exit codes: deriving the camel
spelling instead of reading it, dropping the widget-valued exclusion,
dropping the enum carve-out, breaking the brace matcher and restoring
the literal-space `extends` head each take BOTH checks to exit 1
through the self-test, and exit 0 restored.
